chore: project structure setup
