+++ SwiftUI+MVVM/SwiftUI+MVVM/Coordinators/AuthCoordinator.swift
@@ -1,0 +1,7 @@
+//
+//  AuthCoordinator.swift
+//  SwiftUI+MVVM
+//
+//  Created by Denis Svichkarev on 22/11/24.
+//
+
