+++ SwiftUI+MVVM/SwiftUI+MVVM/Coordinators/CameraCoordinator.swift
@@ -1,0 +1,7 @@
+//
+//  CameraCoordinator.swift
+//  SwiftUI+MVVM
+//
+//  Created by Denis Svichkarev on 22/11/24.
+//
+
