+++ SwiftUI+MVVM/SwiftUI+MVVM/Coordinators/NewsFeedCoordinator.swift
@@ -1,0 +1,7 @@
+//
+//  NewsFeedCoordinator.swift
+//  SwiftUI+MVVM
+//
+//  Created by Denis Svichkarev on 22/11/24.
+//
+
