+++ SwiftUI+MVVM/SwiftUI+MVVM/Core/Features/Auth/Views/SignUpView.swift
@@ -1,0 +1,7 @@
+//
+//  SignUpView.swift
+//  SwiftUI+MVVM
+//
+//  Created by Denis Svichkarev on 22/11/24.
+//
+
