+++ SwiftUI+MVVM/SwiftUI+MVVM/Core/Features/Camera/ViewModels/CameraViewModel.swift
@@ -1,0 +1,7 @@
+//
+//  CameraViewModel.swift
+//  SwiftUI+MVVM
+//
+//  Created by Denis Svichkarev on 22/11/24.
+//
+
