+++ SwiftUI+MVVM/SwiftUI+MVVM/Core/Features/Camera/ViewModels/CreatePostViewModel.swift
@@ -1,0 +1,7 @@
+//
+//  CreatePostViewModel.swift
+//  SwiftUI+MVVM
+//
+//  Created by Denis Svichkarev on 22/11/24.
+//
+
