+++ SwiftUI+MVVM/SwiftUI+MVVM/Core/Features/Camera/Views/CameraView.swift
@@ -1,0 +1,7 @@
+//
+//  CameraView.swift
+//  SwiftUI+MVVM
+//
+//  Created by Denis Svichkarev on 22/11/24.
+//
+
