+++ SwiftUI+MVVM/SwiftUI+MVVM/Core/Features/Camera/Views/CreatePostView.swift
@@ -1,0 +1,7 @@
+//
+//  CreatePostView.swift
+//  SwiftUI+MVVM
+//
+//  Created by Denis Svichkarev on 22/11/24.
+//
+
