+++ SwiftUI+MVVM/SwiftUI+MVVM/Core/Features/Camera/Views/PhotoPreviewView.swift
@@ -1,0 +1,7 @@
+//
+//  PhotoPreviewView.swift
+//  SwiftUI+MVVM
+//
+//  Created by Denis Svichkarev on 22/11/24.
+//
+
