+++ SwiftUI+MVVM/SwiftUI+MVVM/Core/Features/NewsFeed/Views/PostDetailView.swift
@@ -1,0 +1,7 @@
+//
+//  PostDetailView.swift
+//  SwiftUI+MVVM
+//
+//  Created by Denis Svichkarev on 22/11/24.
+//
+
