+++ SwiftUI+MVVM/SwiftUI+MVVM/Core/Features/Profile/ViewModels/ProfileViewModel.swift
@@ -1,0 +1,7 @@
+//
+//  ProfileViewModel.swift
+//  SwiftUI+MVVM
+//
+//  Created by Denis Svichkarev on 22/11/24.
+//
+
