+++ SwiftUI+MVVM/SwiftUI+MVVM/Core/Features/Profile/ViewModels/SettingsViewModel.swift
@@ -1,0 +1,7 @@
+//
+//  SettingsViewModel.swift
+//  SwiftUI+MVVM
+//
+//  Created by Denis Svichkarev on 22/11/24.
+//
+
