+++ SwiftUI+MVVM/SwiftUI+MVVM/Core/Features/Profile/Views/ProfileView.swift
@@ -1,0 +1,7 @@
+//
+//  ProfileView.swift
+//  SwiftUI+MVVM
+//
+//  Created by Denis Svichkarev on 22/11/24.
+//
+
