+++ SwiftUI+MVVM/SwiftUI+MVVM/Core/Models/AppSettings.swift
@@ -1,0 +1,7 @@
+//
+//  AppSettings.swift
+//  SwiftUI+MVVM
+//
+//  Created by Denis Svichkarev on 22/11/24.
+//
+
