+++ SwiftUI+MVVM/SwiftUI+MVVM/Core/Models/Profile.swift
@@ -1,0 +1,7 @@
+//
+//  Profile.swift
+//  SwiftUI+MVVM
+//
+//  Created by Denis Svichkarev on 22/11/24.
+//
+
