+++ SwiftUI+MVVM/SwiftUI+MVVM/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
