+++ SwiftUI+MVVM/SwiftUI+MVVM/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
