+++ SwiftUI+MVVM/SwiftUI+MVVM/Resources/Assets.xcassets/Contents.json
