+++ SwiftUI+MVVM/SwiftUI+MVVM/Resources/MockData/user.json
