+++ SwiftUI+MVVM/SwiftUI+MVVM/Shared/Extensions/View+Extensions.swift
@@ -1,0 +1,7 @@
+//
+//  View+Extensions.swift
+//  SwiftUI+MVVM
+//
+//  Created by Denis Svichkarev on 22/11/24.
+//
+
