+++ SwiftUI+MVVM/SwiftUI+MVVM/Shared/Managers/CameraManager.swift
@@ -1,0 +1,7 @@
+//
+//  CameraManager.swift
+//  SwiftUI+MVVM
+//
+//  Created by Denis Svichkarev on 22/11/24.
+//
+
