+++ SwiftUI+MVVM/SwiftUI+MVVM/Shared/UI/Components/ErrorView.swift
@@ -1,0 +1,7 @@
+//
+//  ErrorView.swift
+//  SwiftUI+MVVM
+//
+//  Created by Denis Svichkarev on 22/11/24.
+//
+
