+++ SwiftUI+MVVM/SwiftUI+MVVM/Shared/UI/Components/LoadingView.swift
@@ -1,0 +1,7 @@
+//
+//  LoadingView.swift
+//  SwiftUI+MVVM
+//
+//  Created by Denis Svichkarev on 22/11/24.
+//
+
